pyproject.toml
[tool.poetry]
name = "fsub"
version = "0.0.1"
description = "File Sharing Telegram Bot"
authors = ["ugorwx"]
license = "gpl-3.0-only"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pyrofork = "^2.3.43"
tgcrypto-pyrofork = "^1.2.6"
uvloop = "^0.19.0"
meval = "^2.5"
python-dotenv = "^1.0.1"
motor = {extras = ["srv"], version = "^3.4.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
